+++ src/components/MainPage/Footer.jsx
