+++ src/components/MainPage/Gift.jsx
