+++ src/components/MainPage/MainPageC.jsx
@@ -1,0 +1,31 @@
+import React from "react";
+import styled from "styled-components";
+import MainContainerComponent from "./MainContainer";
+import Gift from "./Gift";
+import Apparel from "./Apparel";
+import VictoryF from "./VictoryF";
+import LastContainer from "./LastContainer";
+import Ad from "./Ad";
+import Nav from "./Nav";
+import NavFooter from "./NavFooter";
+
+const MainPageContainer = styled.div``;
+
+const MainPageC = () => {
+  return (
+    <>
+      <Nav />
+      <NavFooter />
+      <MainPageContainer>
+        <MainContainerComponent />
+        <Gift />
+        <Apparel />
+        <VictoryF />
+        <LastContainer />
+        <Ad />
+      </MainPageContainer>
+    </>
+  );
+};
+
+export default MainPageC;
