+++ src/components/MainPage/Share.jsx
